working on user cart

+++ Ecomweb/Data/Dto/AddToCartDto.cs
@@ -3,10 +3,10 @@ namespace Ecomweb.Data.Dto;
 public class AddToCartDto
 {
 
-  public long UserId { get; set; }
-  public long? CartId { get; set; }
+  public int UserId { get; set; }
+  public int? CartId { get; set; }
 
-  public long ProductId { get; set; }
+  public int ProductId { get; set; }
 
   public int Quantity { get; set; }
 }
